==> code/01_reddit_api_data_pull.ipynb <==
""" Scrape the r/technews subreddit through the Python Reddit API """

import praw
import pandas as pd
import pickle

# import my Reddit API account info saved in separate .py
from reddit_api_acct import ab_client, ab_secret, ab_agent

reddit = praw.Reddit(client_id = ab_client, 
                     client_secret = ab_secret, 
                     user_agent = ab_agent)


hot_posts = reddit.subreddit('technews').hot(limit=10)
for post in hot_posts:
    print(post.title)

posts = []
ml_subreddit = reddit.subreddit('technews')
for post in ml_subreddit.new(limit=None):  # or .hot
    posts.append([post.title, post.score, post.id, post.subreddit, post.url, post.num_comments, post.selftext, post.created])
posts = pd.DataFrame(posts,columns=['title', 'score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created'])


posts.shape

# even with No Limit, the Reddit API is limited to at max 1K posts

posts.head()

####################

sub = ['technews']  # make a list of subreddits you want to scrape the data from

for s in sub:
    subreddit = reddit.subreddit(s)   # Chosing the subreddit


post_dict = {
    "title" : [],
    "score" : [],
    "id" : [],
    "url" : [],
    "num_comms": [],
    "created" : [],
    "body" : []
}
comments_dict = {
    "comment_id" : [],
    "comment_parent_id" : [],
    "comment_body" : [],
    "comment_link_id" : []
}
    

# ran for top posts on 2/17/2021 

#for submission in subreddit.search(query,sort = "top",limit = 1):
for submission in subreddit.top(limit=None):
    post_dict["title"].append(submission.title)
    post_dict["score"].append(submission.score)
    post_dict["id"].append(submission.id)
    post_dict["url"].append(submission.url)
    post_dict["num_comms"].append(submission.num_comments)
    post_dict["created"].append(submission.created)
    post_dict["body"].append(submission.selftext)

    ##### Accessing comments on the post
    submission.comments.replace_more(limit = None)
    for comment in submission.comments.list():
        comments_dict["comment_id"].append(comment.id)
        comments_dict["comment_parent_id"].append(comment.parent_id)
        comments_dict["comment_body"].append(comment.body)
        comments_dict["comment_link_id"].append(comment.link_id)

post_comments = pd.DataFrame(comments_dict)
post_data = pd.DataFrame(post_dict)
  

post_comments.shape

post_comments.head()

post_data.shape

post_data.head()

# pickle posts and comments

with open('post_data_top.pickle', 'wb') as to_write:
    pickle.dump(post_data, to_write)

    
with open('post_comments_top.pickle', 'wb') as to_write:
    pickle.dump(post_comments, to_write)
