==> code/02_topic_models_sample_posts.ipynb <==
"""
initial exploration of topic modeling of 
r/technews posts pulled through Reddit API
which is limited to max 1k posts
"""

import pandas as pd
import pickle
import datetime

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.decomposition import NMF, LatentDirichletAllocation


with open('post_data_top.pickle', 'rb') as read_file:
    posts_top = pickle.load(read_file)
    
posts_top.shape

##  make the created time stamp from epoch time into date
posts_top['created'] = posts_top.created.astype(int)
posts_top['date_time'] = pd.to_datetime(posts_top['created'], unit='s')

posts_top.head()

min(posts_top.date_time)

#Timestamp('2018-06-23 00:31:27')

# initial topic modeling

## try NMF

vectorizer = CountVectorizer(stop_words = 'english')
title_words = vectorizer.fit_transform(posts_top.title)
title_words.shape

nmf_model = NMF(12)  # tried 5, 7, 10, 12, 15, 20, 25
post_topic = nmf_model.fit_transform(title_words)
post_topic.shape

topic_word = nmf_model.components_
topic_word.shape

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-7:-1]
topic_words = [[words[e] for e in l] for l in t]
topic_words

posts_top['post_topic_nmf'] = post_topic.argmax(axis=1)

posts_top.head()

# Export to Excel to further review posts by NMF topics
posts_top.to_excel('post_topics_sample.xlsx', index=False)

Reviewed the topics assigned to each post title

0. Facebook
1. Tesla, Elon Musk
2. Google
3. SpaceX, NASA
4. variety - commonly of "new"
5. fines, breaches, bad news
6. Apple,  but getting app too and 'apple' the fruit
7. variety
8. about half Twitter, also covid, and other
9. variety
10. Scientists, studies
11. less than half about Amazon


## try LDA

lda_model = LatentDirichletAllocation(n_components=20)
doc_topic = lda_model.fit_transform(topic_word)
doc_topic.shape

words = vectorizer.get_feature_names()
t = lda_model.components_.argsort(axis=1)[:,-1:-7:-1]
topic_words = [[words[e] for e in l] for l in t]
topic_words

# this output is odd that several topics have identical top words
# though LDA is better with larger corpora and this is a sample of 1k posts/sentences

## try TF-IDF vectorizer with NMF

# create the document-term matrix with TF-IDF vectorizer 
cv_tfidf = TfidfVectorizer(stop_words="english")
X_tfidf = cv_tfidf.fit_transform(posts_top.title).toarray()

dt_tfidf = pd.DataFrame(X_tfidf,columns=cv_tfidf.get_feature_names()) 
dt_tfidf

nmf_model = NMF(27)  
post_topic2 = nmf_model.fit_transform(X_tfidf)
post_topic2.shape

words2 = vectorizer.get_feature_names()
t2 = nmf_model.components_.argsort(axis=1)[:,-1:-7:-1]
topic_words = [[words[e] for e in l] for l in t2]
topic_words

## these 12 topics with tf-idf vectorizer worked better than just count vectorizer
## tried 15 topics - looks pretty good
## 20 topics got bitcoin as separate topic
## 25 looks good too
# tried 27 -- it's putting bitcoin with cars

# Need to add more preprocessing after pull all of the data
# vectorizer did lowercase and removed stop words

# try bi-grams 
# entity recognition
# maybe remove the numbers
# stemming/lemmatization
# POS tagging
