==> code/03_PSAW_test.ipynb <==
## Test the PSAW library and Pushshift API to pull all r/technews posts since 2011 

PSAW python library to query copy of all public Reddit posts saved in Pushshift

from psaw import PushshiftAPI
import datetime as dt
import pandas as pd


api=PushshiftAPI()

# The `search_comments` and `search_submissions` methods return generator objects

start_epoch=int(dt.datetime(2011, 1, 1).timestamp())

gen = api.search_submissions(after=start_epoch,
                            subreddit='technews',
                            filter=['url','author', 'title', 'subreddit'],
                            limit=10)

df = pd.DataFrame([thing.d_ for thing in gen])

df.head()

start_epoch=int(dt.datetime(2020, 1, 1).timestamp())

gen2 = api.search_submissions(after=start_epoch,
                            subreddit='technews',
                            limit=10)

df2 = pd.DataFrame([thing.d_ for thing in gen2])

df2.head()

df2.columns

df2[['author', 'domain', 'full_link', 'num_comments', 'score', 'title', 'url', 'created', 'upvote_ratio']].head()

start_epoch=int(dt.datetime(2021, 2, 18).timestamp())

end_epoch=int(dt.datetime(2021, 2, 19).timestamp())

gen3 = api.search_submissions(after=start_epoch, before=end_epoch,
                            subreddit='technews',
                            limit=100)

df3 = pd.DataFrame([thing.d_ for thing in gen3])

df3.head()

df3[['title', 'author', 'domain', 'full_link', 'num_comments', 'score', 'url', 'created', 'upvote_ratio']].head()

df3.to_excel('df3_test.xlsx' ,index=False)

df3.shape

# will need to filter the posts to where not removed
filtered_2 = df3[df3.removed_by_category.isin(['reddit', 'moderator', 'deleted'])]
filtered_2.shape

# filter to where removed_by_category = null
filtered = df3[df3.removed_by_category.isna()]

filtered.shape

filtered
