==> code/04_technews_posts.ipynb <==
"""
Set up pull of r/technews posts 
for 2011 - 2020 
"""

from psaw import PushshiftAPI
import pandas as pd
import datetime as dt

api=PushshiftAPI()

# try to pull 6 months at a time
# keep all fields
# there are less fields in the older years

# with all the fields, will be able to filter out removed posts and many of the ads
# many posts removed by moderator or reddit in the more recent years


start_epoch=int(dt.datetime(2020, 7, 1).timestamp())
end_epoch=int(dt.datetime(2021, 1, 1).timestamp())


gen = api.search_submissions(after=start_epoch, before=end_epoch,
                            subreddit='technews',
                            limit=None)

df = pd.DataFrame([thing.d_ for thing in gen])

df.to_csv('data/technews_' + str(start_epoch) + '_' + str(end_epoch) + '.csv', index=False)

df.shape

df.columns

# 2011 - 473
# 2012 - 91
# 2013 - 85 + 96
# 2014 - 1454 + 1630
# 2015 - 3211 + 2245
# 2016 - 1349 + 4064
# 2017 - 4174 + 3974
# 2018 - 3842 + 5741
# 2019 - 8660 + 7685
# 2020 - 14753 + 13263
