==> code/05_posts_data_cleaning.ipynb <==
"""
combine all the csv files with r/technews posts

filter to remove the deleted posts
check the is_meta and remove those where true/yes
check the archived, quarantine, and is_self columns if any true/yes and remove these posts 

saved filtered dataset to pickle

"""

import glob
import os

import pandas as pd
import pickle

raw_data = pd.DataFrame()

for file_name in glob.glob('data/*.csv'):
    try:
        x = pd.read_csv(file_name, low_memory=False)
        raw_data = pd.concat([raw_data,x],axis=0)
        print('APPENDED: ', file_name)
    except:
        print('ERROR: ', file_name)

raw_data.shape

raw_data.columns

raw_data.head()

raw_data.info()

raw_data.created_utc.dtype

raw_data.describe()

raw_data['date_time'] = pd.to_datetime(raw_data['created_utc'], unit='s')

raw_data.head()

min(raw_data.date_time)

max(raw_data.date_time)

## filter to where removed_by_category = null
raw_data.removed_by_category.value_counts()

f1_df = raw_data[raw_data.removed_by_category.isna()]
f1_df.shape

# examine is_meta as meta posts are posted by subreddit itself

f1_df.is_meta.value_counts()

# export to Excel and review all fields and completeness
f1_df.to_excel('f1_df_review.xlsx', index=False)

# fields to keep


1. author
2. author_fullname
3. created_utc
4. domain --- though this might not be in the oldest data
5. full_link  - link to reddit technews
6. id
7. num_comments
8. permalink  - link starting with r/technews
9. score
10. title
11. url - link to post's article
12. date_time


only recent year has upvote ratio

checked that no true in quarantine, archived, is_self, is_media


# also need to filter where removed_by is blank
# field names changed over the years

f2_df = f1_df[f1_df.removed_by.isna()]
f2_df.shape


f2_df.domain.value_counts()

f2_df.url.nunique()

f2_df.domain.nunique()

f3_df = f2_df[['id','created_utc', 'date_time', 'author', 'author_fullname', 'title', 'url', 
               'domain', 'num_comments', 'score', 'permalink', 'full_link' ]]

f3_df.shape


f3_df.head()

f3_df.info()

with open('technews_posts.pickle', 'wb') as to_write:
    pickle.dump(f3_df, to_write)
