==> code/06_more_data_cleaning.ipynb <==
"""
more data cleaning of r/technews post titles

remove ads 

remove domains from end of post titles

refine entity names to be consistant across post titles
"""

import pickle
import pandas as pd
import numpy as np

with open('technews_posts.pickle', 'rb') as read_file:
    posts = pickle.load(read_file)
    
posts.shape

posts.columns

# after review of posts 
# saw couple thousand of printer help ads need removed
# list of authors who posted these ads

ad_authors =['rakeshrcs', 'isabellajac', 'martinhenry821', 'mikeholder609', 
             'rachelzoe58', 'rakeshrcs', 'heskettsheila769', 'mariyajhones902', 
             'jasonroy9540', 'jhonbrother93', 'jeniferberoy', 'clakejhon789', 
             'Jordnbob', 'epsonsupport', 'davidjhonson584', 'clarkstward1919', 
             'canonprintersupport9', 'barendamorre', 'anthony_scoble', 'alberthuddles', 
             'joeclake35', 'joejim748', 'lindaamartin253', 'lxiswift82', 'mariyajhones902',
             'martinhenry821', 'miguelabad', 'mikeholder609', 'MilenaEllen', 
             'printerstelecomm', 'qbnumber', 'rachelzoe58', 'renitaa1', 'robertamy284', 
             'shyrasimon26', 'Smithmkt', 'taniacasas', 'johnbradny', 'johnsmith44210',
             'jordnbob', 'jp_todaro', 'js637209', 'juliemiler151', 'Junaid0005', 
             'justmestunner', 'katrinaa2', 'kawtarkist', 'kinfoprinters', 'Carindermalhi',
             'kloudconnectors','KQNk62', 'ksourav', 'Kvsc', 'latonya1', 'up525']


len(posts[posts['author'].isin(ad_authors)])

f1_posts = posts[~posts['author'].isin(ad_authors)]

f1_posts.shape

# more printer ads to remove by domain
# domain self.technews is posts from subreddit moderators to channel members

ad_domain = ['hp-support-us.blogspot.com', '123printersupport.com',  '123hp.us', 'self.technews']

len(f1_posts[f1_posts['domain'].isin(ad_domain)])

f2_posts = f1_posts[~f1_posts['domain'].isin(ad_domain)]

f2_posts.shape

# over a thousand posts with this non-content 

len(f2_posts[f2_posts['title'] == 'Pinned to News and Press on Pinterest'])


# remove 
f3_posts = f2_posts[f2_posts['title'] != 'Pinned to News and Press on Pinterest']
f3_posts.shape

###################

# many posts end with the domain of the news article, remove domains from end of post title

len(f3_posts[f3_posts['title'].str[-10:] =='TechFeasta'])

f3_posts.loc[:, 'refine_title'] = np.where(f3_posts.loc[:,'title'].str[-11:] ==' TechFeasta', 
         f3_posts['title'].str[:-11], 
         f3_posts['title'])

f3_posts.shape

len(f3_posts[f3_posts['title'].str[-6:] ==' WIRED'])

f3_posts.loc[:,'refine_title'] = np.where(f3_posts.loc[:,'title'].str[-6:] ==' WIRED', 
         f3_posts['title'].str[:-6], 
         f3_posts['refine_title'])

f3_posts.shape

len(f3_posts[f3_posts['title'].str[-16:] ==' Technology News'])

f3_posts.loc[:,'refine_title'] = np.where(f3_posts.loc[:,'title'].str[-16:] ==' Technology News', 
         f3_posts['title'].str[:-16], 
         f3_posts['refine_title'])

len(f3_posts[f3_posts['title'].str[-8:] ==' insopra'])

f3_posts.loc[:,'refine_title'] = np.where(f3_posts.loc[:,'title'].str[-8:] ==' insopra', 
         f3_posts['title'].str[:-8], 
         f3_posts['refine_title'])

len(f3_posts[f3_posts['title'].str[-15:] =='DigiTech Engine'])

f3_posts.loc[:,'refine_title'] = np.where(f3_posts.loc[:,'title'].str[-15:] =='DigiTech Engine', 
         f3_posts['title'].str[:-15], 
         f3_posts['refine_title'])

len(f3_posts[f3_posts['title'].str[-24:] ==" Science &amp; Tech News"])

f3_posts.loc[:,'refine_title'] = np.where(f3_posts.loc[:,'title'].str[-24:] ==" Science &amp; Tech News", 
         f3_posts['title'].str[:-24], 
         f3_posts['refine_title'])

##############
# refine company/entity names where inconsistant

f3_posts.loc[:,'refine_title'] = f3_posts.loc[:,'refine_title'].str.replace("AT&amp;T", 'ATT')
f3_posts.loc[:,'refine_title'] = f3_posts.loc[:,'refine_title'].str.replace("N.S.A.", 'NSA')
f3_posts.loc[:,'refine_title'] = f3_posts.loc[:,'refine_title'].str.replace('Tik Tok', 'TikTok')

# review output
#3_posts.to_excel('f3_posts_review.xlsx', index=False)

f3_posts.columns

with open('filtered_technews_posts.pickle', 'wb') as to_write:
    pickle.dump(f3_posts, to_write)
