==> code/07_title_text_preprocessing.ipynb <==
"""
preprocess post title text of all r/technews posts 2011-2020
"""

import pandas as pd
import numpy as np
import pickle
import datetime

import re
import string

import nltk
from nltk.chunk import ne_chunk
from nltk.tag import pos_tag
from nltk.tokenize import word_tokenize

with open('filtered_technews_posts.pickle', 'rb') as read_file:
    posts = pickle.load(read_file)
    
posts.shape

posts.head()

string.punctuation

# Text preprocessing steps - remove numbers and punctuation

alphanumeric = lambda x: re.sub('\w*\d\w*', ' ', x)
punc_remove = lambda x: re.sub('[%s]' % re.escape(string.punctuation), ' ', x)

posts['preproc_title'] = posts['refine_title'].map(alphanumeric).map(punc_remove)
posts.head()

## stem the title text

#lemma = nltk.stem.WordNetLemmatizer()  ## this only made some of the plural words singular

from nltk.stem.snowball import SnowballStemmer

stemmer = SnowballStemmer("english")

posts['unstemmed'] = posts.preproc_title.str.split()
#posts['stem_title'] = posts.unstemmed.apply(lambda x: [lemma.lemmatize(y) for y in x])

posts['stem_title'] = posts.unstemmed.apply(lambda x: [stemmer.stem(y) for y in x])

posts['processed_title'] = posts['stem_title'].apply(lambda x: ' '.join(x))

posts.head()

# will try topic modeling with only nouns
# POS tag to keep nouns

def keep_nouns(title_text):
    tokens = word_tokenize(title_text) 
    tags = pos_tag(tokens) # this labels each word as a part of speech
    nouns = [word for word,pos in tags if (pos == 'NN' or pos == 'NNP' or pos == 'NNS' or pos == 'NNPS')]
    return ' '.join(nouns)

posts['title_nouns'] = posts['preproc_title'].apply(keep_nouns)
posts.head()

with open('preprocessed_posts.pickle', 'wb') as to_write:
    pickle.dump(posts, to_write)
