==> code/08_topic_modeling.ipynb <==
"""
topic modeling of r/technews posts
"""

import pandas as pd
import numpy as np
import pickle
import datetime

import nltk
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.decomposition import NMF, LatentDirichletAllocation

with open('preprocessed_posts.pickle', 'rb') as read_file:
    posts = pickle.load(read_file)
    
posts.shape

posts.head()

# add to list of stop words

stopwords = nltk.corpus.stopwords.words('english')
stopwords.extend(['new', 'news', 'tech', '&amp;', 'amp', 'million', 'billion'] )


### initial topic model

vectorizer = CountVectorizer(stop_words = stopwords, ngram_range=(1,2))
title_words = vectorizer.fit_transform(posts.processed_title)
title_words.shape

nmf_model = NMF(25)  # tried 5, 7, 10, 12, 15, 20, 25, 30
post_topic = nmf_model.fit_transform(title_words)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-7:-1]
topic_words = [[words[e] for e in l] for l in t]
topic_words

# with 20 NMF above, ai in two topics
# still in ai in 2 topics with 25 topics

## try topic modeling with title_nouns

vectorizer = CountVectorizer(stop_words = stopwords, ngram_range=(1,2))
title_words = vectorizer.fit_transform(posts.title_nouns)
title_words.shape

nmf_model = NMF(20)  # tried 5, 7, 10, 12, 15, 17, 20, 25, 27, 30
post_topic = nmf_model.fit_transform(title_words)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-7:-1]
topic_words = [[words[e] for e in l] for l in t]
topic_words

# 20 topics got ai in a separate topic, tesla in it's own topic
# most of these 20 can be easily labeled

# the CountVectorizer + NMF yielded pretty seaparate and easily labeled topics
# with AI and machine learning separate
# a couple coule be combined

posts['post_topic_nmf'] = post_topic.argmax(axis=1)

post_info = posts.reset_index()

# add in the topic components

components = pd.DataFrame(post_topic.round(5),
             index = posts.title_nouns,
             columns = ["C_1","C_2", "C_3", "C_4", "C_5", "C_6", "C_7", "C_8", "C_9", 
                       "C_10", "C_11", "C_12", "C_13", "C_14", "C_15", "C_16", "C_17",
                       "C_18", "C_19", "C_20"])

c_info = components.reset_index()

c_info.head()

c_info.shape

post_info.shape

# add the topic components to the full df

all_df = pd.concat([post_info, c_info], axis=1)
all_df.shape
        

# the CountVectorizer + NMF yielded pretty separate and easily labeled topics
# will examine if any benefit with clustering topic components

with open('cv_nmf_topics_posts.pickle', 'wb') as to_write:
    pickle.dump(all_df, to_write)

####################

# CountVectorizer + LDA

vectorizer = CountVectorizer(stop_words = stopwords, ngram_range=(1,2))
title_words = vectorizer.fit_transform(posts.processed_title)

lda_model = LatentDirichletAllocation(n_components=20)
doc_topic = lda_model.fit_transform(title_words)
doc_topic.shape

words = vectorizer.get_feature_names()
t = lda_model.components_.argsort(axis=1)[:,-1:-7:-1]
topic_words = [[words[e] for e in l] for l in t]
topic_words

# google in almost every topic with title_nouns input
# with processed_title still has some key words in multiple topics

###################

### try tfidf vectorizer

# create the document-term matrix with TF-IDF vectorizer 
cv_tfidf = TfidfVectorizer(stop_words=stopwords)
X_tfidf = cv_tfidf.fit_transform(posts.title_nouns).toarray()


nmf_model = NMF(27)  
post_topic2 = nmf_model.fit_transform(X_tfidf)
post_topic2.shape

words2 = vectorizer.get_feature_names()
t2 = nmf_model.components_.argsort(axis=1)[:,-1:-7:-1]
topic_words = [[words[e] for e in l] for l in t2]
topic_words
