==> code/12_refine_kmeans_topic_components.ipynb <==
""" k-means of NMF topic components"""

import pickle
import pandas as pd
import numpy as np


# see https://ipython.readthedocs.io/en/stable/interactive/magics.html
%pylab inline

# sets backend to render higher res images
%config InlineBackend.figure_formats = ['retina']

from sklearn.cluster import KMeans
import seaborn as sns

with open('cv_nmf_topics_posts_FINAL.pickle', 'rb') as read_file:
    posts = pickle.load(read_file)
    
posts.shape

posts.columns

X = np.array(posts[['C_0', 'C_1', 'C_2', 'C_3', 'C_4', 'C_5', 'C_6', 'C_7', 'C_8', 'C_9',
       'C_10', 'C_11', 'C_12', 'C_13', 'C_14']])

# helper function that allows us to display data in 2 dimensions an highlights the clusters
def display_cluster(X,km=[],num_clusters=0):
    color = 'brgcmyk'
    alpha = 0.5
    s = 20
    if num_clusters == 0:
        plt.scatter(X[:,0],X[:,1],c = color[0],alpha = alpha,s = s)
    else:
        for i in range(num_clusters):
            plt.scatter(X[km.labels_==i,0],X[km.labels_==i,1],c = color[i],alpha = alpha,s=s)
            plt.scatter(km.cluster_centers_[i][0],km.cluster_centers_[i][1],c = color[i], marker = 'x', s = 100)

num_clusters = 12
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=5) # n_init, number of times the K-mean algorithm will run
km.fit(X)


#display_cluster(X,km,num_clusters)

clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])
#clusters


len(clusters)

clusters.head()

order_centroids = km.cluster_centers_.argsort()[:, ::-1] 
order_centroids[0]

# add k-means clusters to df

posts_clusters = pd.concat([posts, clusters], axis=1)


posts_clusters.shape

posts_clusters.columns

subset = posts_clusters[['title','post_topic_nmf', 'cluster']]

## examined for each cluster #

subset[subset['cluster']==11]
