==> code/13_sentiment_analysis_test.ipynb <==
"""
comparing TextBlob with Vader sentiment analysis
of r/technews comments
"""

import pickle
import pandas as pd
import numpy as np

from textblob import TextBlob

# TextBlob and Vader sentiment analysis tools were trained on internet posts, 
# and should work well for the subreddit comments

"""The sentiment property returns a named tuple of the form Sentiment(polarity, subjectivity).
The polarity score is a float within the range [-1.0, 1.0]. The subjectivity is a float 
within the range [0.0, 1.0] where 0.0 is very objective and 1.0 is very subjective."""

with open('post_comments_top.pickle', 'rb') as read_file:
    comments = pickle.load(read_file)
    
comments.shape

comments.head()

comments['sentiment_polarity'] = comments['comment_body'].apply(lambda x: TextBlob(x).sentiment.polarity)

comments['sentiment_subjectivity'] = comments['comment_body'].apply(lambda x: TextBlob(x).sentiment.subjectivity)

comments.head()

min(comments.sentiment_polarity)

comments[comments['sentiment_polarity']==-1].head()

comments.iloc[289, 2]

comments.iloc[878, 2]

comments.iloc[1355, 2]

max(comments.sentiment_polarity)

comments[comments['sentiment_polarity']==1].head()

comments.iloc[113, 2]

comments.iloc[144, 2]

comments.iloc[298, 2]

#################

#compare to Vader

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer 

sid_obj = SentimentIntensityAnalyzer()
sentiment = []
for text in comments.comment_body:
    sentiment.append(sid_obj.polarity_scores(text))

len(sentiment)

comments.shape

vader_results = pd.concat([comments,pd.DataFrame(sentiment)], axis=1)

vader_results.head()

#the "compound" score, ranging from -1 (most neg) to 1 (most pos) would provide a single measure of polarity.

"""
The Compound score is a metric that calculates the sum of all the 
lexicon ratings which have been normalized between -1(most extreme negative) 
and +1 (most extreme positive).

positive sentiment : (compound score >= 0.05)
neutral sentiment : (compound score > -0.05) and (compound score < 0.05)
negative sentiment : (compound score <= -0.05)"""
