==> code/14_technews_comments.ipynb <==
"""
Set up pull of r/technews comments 
for 2011 - 2020 
"""

from psaw import PushshiftAPI
import pandas as pd
import datetime as dt

api=PushshiftAPI()

# try to pull 6 months at a time
# keep all fields
# there are less fields in the older years

# with all the fields, will be able to filter out removed posts and many of the ads
# many posts removed by moderator or reddit in the more recent years


start_epoch=int(dt.datetime(2020, 10, 1).timestamp())
end_epoch=int(dt.datetime(2021, 1, 1).timestamp())


gen = api.search_comments(after=start_epoch, before=end_epoch,
                            subreddit='technews',
                            limit=None)

df = pd.DataFrame([thing.d_ for thing in gen])

df.to_csv('data/COMMENTS_technews_' + str(start_epoch) + '_' + str(end_epoch) + '.csv', index=False)

df.shape

df.columns

# COMMENTS PER YEAR

# 2011 - 123
# 2012 - 45
# 2013 - 37
# 2014 - 6553
# 2015 - 4007 + 2445
# 2016 - 1493 + 1633
# 2017 - 2057 + 2156
# 2018 - 5030 + 24468
# 2019 - 32309 + 44847
# 2020 - 29417 + 44284 + 43003 + 55121

"""
For each comment, the parent can be either a submission or another comment. 

the t# prefixes refer to the type of thing in the reddit data structure. 

type prefixes
t1_	Comment
t2_	Account
t3_	Link
t4_	Message
t5_	Subreddit
t6_	Award


"""

df[['id', 'link_id', 'parent_id']].head()

# id is the comment id
# match link_id to submission/post id
# use parent_id for network analysis, as will capture one commenter responding to another commenter 

#where user_removed=True - the user deleted the comment
