==> code/15_comments_data_cleaning.ipynb <==
"""
combine all the csv files with r/technews comments

filter to remove the deleted comments

check other fields for indicators of removing any ads or comments by bots

save filtered dataset to pickle

"""

import glob
import os

import pandas as pd
import numpy as np
import pickle

x = pd.read_csv('data/COMMENTS_technews_1561953600_1577854800.csv', low_memory=False)
x.head()

x.columns

raw_data = pd.DataFrame()

for file_name in glob.glob('data/COMMENTS_*.csv'):
    try:
        x = pd.read_csv(file_name, low_memory=False)
        raw_data = pd.concat([raw_data,x],axis=0)
        print('APPENDED: ', file_name)
    except:
        print('ERROR: ', file_name)

raw_data.shape

raw_data.columns

# more fields in more recent years

raw_data.head()

raw_data.info()

raw_data.describe()

# export to Excel and review all fields and completeness
raw_data.to_excel('comments_df_review.xlsx', index=False)

# after review of all and field completeness

# filter from sentiment analysis to remove where author is
#AutoModerator
#PineappleMeister
#babar77
# these are the subreddit moderators commenting on reason for removing post


# also remove where:
# stickied=TRUE
# mod_removed = 'TRUE'
# user_removed = 'TRUE'
# body = [removed]
# body = [deleted]


# will KEEP where author = [deleted] and the comment body is still there
# occurs where user that cancelled reddit account

raw_data.shape

len(raw_data[raw_data['author'].isin(['AutoModerator','PineappleMeister', 'babar77'] )])

f1_comments = raw_data[~raw_data['author'].isin(['AutoModerator','PineappleMeister', 'babar77'])]
f1_comments.shape                                    

len(f1_comments[f1_comments['stickied'] ==True])


f2_comments = f1_comments[f1_comments['stickied']!=True]
f2_comments.shape

len(f2_comments[f2_comments['mod_removed'] ==True])

f3_comments = f2_comments[f2_comments['mod_removed'] !=True]
f3_comments.shape

len(f3_comments[f3_comments['user_removed'] == True])

f4_comments = f3_comments[f3_comments['user_removed'] != True]
f4_comments.shape

len(f4_comments[f4_comments['body'] =='[removed]'])

f5_comments = f4_comments[f4_comments['body'] !='[removed]']
f5_comments.shape

f6_comments = f5_comments[f5_comments['body'] !='[deleted]']
f6_comments.shape

f7_comments = f6_comments[~f6_comments['body'].isin([np.nan])]

f8_comments = f7_comments.reset_index(drop=True)

f8_comments.shape

# comment fields to keep


1. author
2. author_created_utc
3. body
4. created_utc
5. id
6. link_id
7. parent_id
8. permalink
9. retrieved_on
10. score
11. created
12. subreddit
13. updated_utc





comments_fields = f8_comments[['author', 'author_created_utc', 'body', 'created_utc', 'id', 
                              'link_id', 'parent_id', 'permalink', 'retrieved_on', 
                              'score', 'created', 'subreddit', 'updated_utc']]

comments_fields.shape


comments_fields.head()

comments_fields.info()

with open('technews_comments.pickle', 'wb') as to_write:
    pickle.dump(comments_fields, to_write)
