==> code/19_visualize_topics.ipynb <==
"""
Add label topics

aggregate number of posts by topic 

export to excel to graph
"""

import pickle
import pandas as pd
import numpy as np
import datetime as dt

with open('cv_nmf_topics_posts_FINAL.pickle', 'rb') as read_file:
    topics_df = pickle.load(read_file)
    
topics_df.shape

topics_df.head()

topics_df['YEAR_MON'] = topics_df['date_time'].dt.strftime('%Y-%m')
topics_df.head()

topics_df['QUARTER'] = topics_df['date_time'].dt.quarter.astype(str)
topics_df['YEAR']= topics_df['date_time'].dt.year
topics_df['YEAR_QTR'] = topics_df['date_time'].dt.strftime('%Y') + 'Q' + topics_df['QUARTER']
topics_df.head()

count_posts = topics_df[['title', 'post_topic_nmf', "YEAR_MON", 'YEAR', 'YEAR_QTR']].groupby(['post_topic_nmf', 'YEAR_MON', 'YEAR', 'YEAR_QTR']).count().reset_index()
count_posts.head()

# will combine the 2 App topics
# also combine the 3 phone topics

count_posts['topic_number'] = np.where(count_posts['post_topic_nmf'] == 4, 0, count_posts['post_topic_nmf'])

count_posts['topic_number'] = np.where(count_posts['post_topic_nmf'].isin([6,8]), 2, count_posts['topic_number'])

count_posts.topic_number.nunique()

count_posts[count_posts.post_topic_nmf == 4].head()

count_posts[count_posts.post_topic_nmf == 6].head()

count_posts[count_posts.post_topic_nmf == 8].head()

## aggregate for the combined topics

posts_agg = count_posts[["YEAR_MON", 'YEAR', 'YEAR_QTR', 'topic_number', 'title']].groupby(['YEAR_MON', 'YEAR', 'YEAR_QTR', 'topic_number']).agg({'title':np.sum}).reset_index()
posts_agg

## Topic Labels

0. Apps
1. Data Breaches
2. Phones
3. Computers
4. MERGED with 0
5. Security 
6. MERGED with 2
7. International
8. MERGED with 2
9. AI
10. Internet
11. Printers
12. Cameras
13. Customer Support
14. Software Development

# put labels in df

labels_dict = {0:'Apps', 1:'Data Breaches', 2:'Phones', 3:'Computers', 5:'Security', 
              7:'International', 9:'Artifical Intelligence', 10:'Internet', 11:'Printers', 12:'Cameras', 
              13:'Customer Support', 14:'Software Development'}

labels_df = pd.DataFrame.from_dict(labels_dict, orient='index', columns=['topic_label'])

labels_df

merge_df = pd.merge(posts_agg, labels_df, how='left', left_on='topic_number', right_index=True)

merge_df.head()

# pivot so one row for each month, topics to their own column

pivot_df = merge_df.pivot_table(index=['YEAR_MON', 'YEAR', 'YEAR_QTR'], columns='topic_label', values='title', fill_value=0).reset_index()
pivot_df.head()

pivot_df.to_excel('topic_counts_by_month.xlsx', index=False)

## aggregate by quarter
quarter_agg = merge_df[['YEAR_QTR', 'topic_number', 'topic_label', 'title']].groupby(['YEAR_QTR', 'topic_number','topic_label']).agg({'title':np.sum}).reset_index()
quarter_agg.head()

pivot_df.to_excel('topic_counts_by_quarter.xlsx', index=False)

######
# adjust pivot for topic labels in row and time in columns 
pivot_qtr = quarter_agg.pivot_table(index=['topic_label'], columns='YEAR_QTR', values='title', fill_value=0).reset_index()
pivot_qtr.head()

pivot_qtr.to_excel('topic_counts_by_label_qtr.xlsx', index=False)

# by year
year_agg =merge_df[['YEAR', 'topic_number', 'topic_label', 'title']].groupby(['YEAR', 'topic_number','topic_label']).agg({'title':np.sum}).reset_index()

year_agg.head()

pivot_df.to_excel('topic_counts_by_year.xlsx', index=False)

# aggregate by topic

topic_agg =merge_df[['topic_label', 'title']].groupby(['topic_label']).agg({'title':np.sum}).reset_index()

topic_agg

sum(topic_agg.title)

topic_agg.to_excel('by_topic_counts.xlsx', index=False)
